refactor: loader
