+++ lib/loader.js
@@ -2,15 +2,23 @@
   MIT License http://www.opensource.org/licenses/mit-license.php
   Author Tobias Koppers @sokra
 */
+const postcss = require('postcss');
+const localByDefault = require('postcss-modules-local-by-default');
+const extractImports = require('postcss-modules-extract-imports');
+const modulesScope = require('postcss-modules-scope');
+const modulesValues = require('postcss-modules-values');
 const loaderUtils = require('loader-utils');
 
-const processCss = require('./processCss');
+const { importParser, icssParser, urlParser } = require('./plugins');
 const {
+  getLocalIdent,
   getImportPrefix,
   placeholderImportItemReplacer,
   compileExports,
   placholderRegExps,
 } = require('./utils');
+const Warning = require('./Warning');
+const CssSyntaxError = require('./CssSyntaxError');
 
 module.exports = function loader(content, map) {
   const callback = this.async();
@@ -35,27 +43,92 @@ module.exports = function loader(content, map) {
   }
   /* eslint-enable no-param-reassign */
 
-  processCss(
-    content,
-    map,
-    {
-      loaderContext: this,
-      loaderOptions: options,
-      sourceMap,
-    },
-    (err, result) => {
-      if (err) {
-        return callback(err);
-      }
+  const loaderContext = this;
+  const localIdentName = options.localIdentName || '[hash:base64]';
+  const customGetLocalIdent = options.getLocalIdent || getLocalIdent;
+
+  const parserOptions = {
+    url: options.url !== false,
+    import: options.import !== false,
+  };
+
+  const plugins = [
+    modulesValues,
+    localByDefault({
+      mode: options.modules ? 'local' : 'global',
+      rewriteUrl(global, url) {
+        if (parserOptions.url) {
+          // eslint-disable-next-line no-param-reassign
+          url = url.trim();
+
+          if (
+            !url.replace(/\s/g, '').length ||
+            !loaderUtils.isUrlRequest(url)
+          ) {
+            return url;
+          }
+          if (global) {
+            return loaderUtils.urlToRequest(url);
+          }
+        }
+        return url;
+      },
+    }),
+    extractImports(),
+    modulesScope({
+      generateScopedName: function generateScopedName(exportName) {
+        return customGetLocalIdent(loaderContext, localIdentName, exportName, {
+          regExp: options.localIdentRegExp,
+          hashPrefix: options.hashPrefix || '',
+          context: options.context,
+        });
+      },
+    }),
+  ];
+
+  if (options.import !== false) {
+    plugins.push(importParser(parserOptions));
+  }
+
+  if (options.url !== false) {
+    plugins.push(urlParser(parserOptions));
+  }
+
+  plugins.push(icssParser(parserOptions));
+
+  postcss(plugins)
+    .process(content, {
+      // we need a prefix to avoid path rewriting of PostCSS
+      from: `/css-loader!${loaderUtils
+        .getRemainingRequest(this)
+        .split('!')
+        .pop()}`,
+      to: loaderUtils
+        .getCurrentRequest(this)
+        .split('!')
+        .pop(),
+      map: options.sourceMap
+        ? {
+            prev: map,
+            sourcesContent: true,
+            inline: false,
+            annotation: false,
+          }
+        : null,
+    })
+    .then((result) => {
+      result
+        .warnings()
+        .forEach((warning) => this.emitWarning(new Warning(warning)));
 
       // for importing CSS
       const importUrlPrefix = getImportPrefix(this, options);
 
       let exportJs = compileExports(
-        result.exports,
+        parserOptions.exports,
         placeholderImportItemReplacer(
           this,
-          result.importItems,
+          parserOptions.importItems,
           importUrlPrefix,
           options.exportOnlyLocals
         ),
@@ -70,10 +143,10 @@ module.exports = function loader(content, map) {
         return callback(null, exportJs);
       }
 
-      let cssAsString = JSON.stringify(result.source);
+      let cssAsString = JSON.stringify(result.css);
 
       const alreadyImported = {};
-      const importJs = result.importItems
+      const importJs = parserOptions.importItems
         .filter((imp) => {
           if (!imp.media) {
             if (alreadyImported[imp.url]) {
@@ -104,7 +177,11 @@ module.exports = function loader(content, map) {
 
       cssAsString = cssAsString.replace(
         placholderRegExps.importItemG,
-        placeholderImportItemReplacer(this, result.importItems, importUrlPrefix)
+        placeholderImportItemReplacer(
+          this,
+          parserOptions.importItems,
+          importUrlPrefix
+        )
       );
 
       // helper for ensuring valid CSS strings from requires
@@ -112,8 +189,8 @@ module.exports = function loader(content, map) {
 
       if (
         options.url !== false &&
-        result.urlItems &&
-        result.urlItems.length > 0
+        parserOptions.urlItems &&
+        parserOptions.urlItems.length > 0
       ) {
         urlEscapeHelper = `var escape = require(${loaderUtils.stringifyRequest(
           this,
@@ -125,7 +202,7 @@ module.exports = function loader(content, map) {
           (item) => {
             const match = placholderRegExps.urlItem.exec(item);
             let idx = +match[1];
-            const urlItem = result.urlItems[idx];
+            const urlItem = parserOptions.urlItems[idx];
             const { url } = urlItem;
 
             idx = url.indexOf('?#');
@@ -164,7 +241,8 @@ module.exports = function loader(content, map) {
       if (sourceMap && result.map) {
         /* eslint-disable no-param-reassign */
         // Add a SourceMap
-        ({ map } = result);
+        map = result.map.toJSON();
+
         if (map.sources) {
           map.sources = map.sources.map(
             (source) =>
@@ -176,12 +254,14 @@ module.exports = function loader(content, map) {
           );
           map.sourceRoot = '';
         }
+
         map.file = map.file
           .split('!')
           .pop()
           .replace(/\\/g, '/');
         map = JSON.stringify(map);
         /* eslint-enable no-param-reassign */
+
         moduleJs = `exports.push([module.id, ${cssAsString}, "", ${map}]);`;
       } else {
         moduleJs = `exports.push([module.id, ${cssAsString}, ""]);`;
@@ -198,6 +278,10 @@ module.exports = function loader(content, map) {
           `// module\n${moduleJs}\n\n` +
           `// exports\n${exportJs}`
       );
-    }
-  );
+    })
+    .catch((error) => {
+      callback(
+        error.name === 'CssSyntaxError' ? new CssSyntaxError(error) : error
+      );
+    });
 };
